+++ src/main/java/sky/pro/SkyDreamTeam/AnimalService/model/Client.java
@@ -1,4 +1,4 @@
-package sky.pro.SkyDreamTeam.AnimalService.Model;
+package sky.pro.SkyDreamTeam.AnimalService.model;
 
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
@@ -13,6 +13,6 @@ public class Client {
     private String name;
     private String phone;
     private String address;
-
+    //private Enum lastMenu;Если пользователь уже обращался к боту ранее, то новое обращение начинается с выбора запроса, с которым пришел пользователь.
     private boolean isVolunteer;
 }
