+++ src/main/java/sky/pro/SkyDreamTeam/AnimalService/model/Information.java
@@ -1,18 +1,14 @@
-package sky.pro.SkyDreamTeam.AnimalService.Model;
-
+package sky.pro.SkyDreamTeam.AnimalService.model;
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
 @Entity
-public class Volunteer {
+public class Information {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
-    private Long chatId;
-    private String name;
-    private String phone;
-    private String address;
+    private String question;
+    private String answer;
 
-    private boolean isVolunteer;
 }
