+++ src/main/java/sky/pro/SkyDreamTeam/AnimalService/model/Volunteer.java
@@ -1,14 +1,18 @@
-package sky.pro.SkyDreamTeam.AnimalService.Model;
+package sky.pro.SkyDreamTeam.AnimalService.model;
+
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
 @Entity
-public class Information {
+public class Volunteer {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
-    private String question;
-    private String answer;
+    private Long chatId;
+    private String name;
+    private String phone;
+    private String address;
 
+    private boolean isVolunteer;
 }
