+++ src/main/java/sky/pro/SkyDreamTeam/AnimalService/repository/ClientRepository.java
@@ -1,7 +1,7 @@
 package sky.pro.SkyDreamTeam.AnimalService.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import sky.pro.SkyDreamTeam.AnimalService.Model.Client;
+import sky.pro.SkyDreamTeam.AnimalService.model.Client;
 
 public interface ClientRepository extends JpaRepository<Client, Long> {
 }
