+++ src/main/java/sky/pro/SkyDreamTeam/AnimalService/repository/InformationRepository.java
@@ -1,7 +1,7 @@
 package sky.pro.SkyDreamTeam.AnimalService.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import sky.pro.SkyDreamTeam.AnimalService.Model.Information;
+import sky.pro.SkyDreamTeam.AnimalService.model.Information;
 
 public interface InformationRepository extends JpaRepository<Information, Long> {
 }
