+++ src/main/java/sky/pro/SkyDreamTeam/AnimalService/repository/PetRepository.java
@@ -1,7 +1,7 @@
 package sky.pro.SkyDreamTeam.AnimalService.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import sky.pro.SkyDreamTeam.AnimalService.Model.Pet;
+import sky.pro.SkyDreamTeam.AnimalService.model.Pet;
 
 public interface PetRepository extends JpaRepository<Pet, Long> {
 }
