+++ src/main/java/sky/pro/SkyDreamTeam/AnimalService/repository/RecordRepository.java
@@ -1,7 +1,7 @@
 package sky.pro.SkyDreamTeam.AnimalService.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import sky.pro.SkyDreamTeam.AnimalService.Model.Record;
+import sky.pro.SkyDreamTeam.AnimalService.model.Record;
 
 public interface RecordRepository extends JpaRepository<Record, Long> {
 }
