+++ src/main/java/sky/pro/SkyDreamTeam/AnimalService/repository/VolunteerRepository.java
@@ -1,7 +1,7 @@
 package sky.pro.SkyDreamTeam.AnimalService.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import sky.pro.SkyDreamTeam.AnimalService.Model.Volunteer;
+import sky.pro.SkyDreamTeam.AnimalService.model.Volunteer;
 
 public interface VolunteerRepository extends JpaRepository<Volunteer, Long> {
 }
